requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
